backend/requirements.txt
# Core API & Server
fastapi==0.110.1
uvicorn[standard]==0.25.0
pydantic[email]==2.12.5
python-multipart==0.0.22

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Security & Authentication
PyJWT==2.11.0
python-jose[cryptography]==3.5.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4

# Integrations
razorpay==2.0.0
requests==2.32.5
aiohttp==3.13.3

# Utilities
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
pyyaml==6.0.3

# Note: emergentintegrations is assumed to be provided by the platform environment
